+++ ios-swiftui/samples/host/PlanetSampleHost/PlanetSampleHostApp.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+import PlanetComponentsSamples
+
+@main
+struct PlanetSampleHostApp: App {
+    var body: some Scene {
+        WindowGroup {
+            BasicControlsSampleView()
+        }
+    }
+}